Allow streaming responses 🏞
This allows services to stream responses back to clients, and clients to process such responses gradually. If a service wishes to return a streaming response, it sets response.Body to a new “Steamer” object to which it can then write. This interface does not provide buffering; if services need this they can layer a buffered (bufio) writer on top.