Don't chunk small request bodies 🎁
For small requests it is not desirable to chunk the body as it means intermediary proxies (like linkerd) cannot retry the request transparently. Go's logic for deciding when to chunk a request is subtle and a bit magic[1], but this prevents chunking behaviour when Encode() is called before sending the response off.

[1] https://github.com/golang/go/blob/26708439ecb4dc923149b17eb522df46dcc2fb23/src/net/http/transfer.go#L150-L186