Fix leak when infinite contexts are used to make requests 🚰
Previously when a caller used an infinite context (eg. context.Background()) to make a request, this would cause a resource leak as Typhon was waiting for the context to be cancelled to Close() the body. This change ensures request bodies are additionally closed when they are fully read, and adds tests to ensure no goroutines are leaked.