Automatically close request/response bodies when they are Decoded
The convenience method Decode() uses a JSON decoder internally. This can leave the end of the body, which typically consists of a terminating newline character, unread. This is undesirable: Decode() is a one-shot operation, and is not suitable for calling repeatedly to read multiple values anyway. To mitigate leaving the underlying reader "hanging" in this situation, the body is now closed within these methods.