Replace Router internals ⚡️
This replaces the internals of Router to use regular expressions instead of wrapping labstack/echo's router. This means the implementation is much simpler, correct, and flexible, at the cost of a small performance hit. This penalty is very small, and in the grand scheme of all the more expensive things a handlr will typically do, is tiny. This follows engineering principle[1] #8.

Fixes #104.

[1]: https://monzo.com/blog/2018/06/29/engineering-principles/