Simplify Content-Length logic on Request.Write/Response.Write
The Content-Length on Requests and Responses is now initialised to zero (instead of -1 which indicates that it is unknown), and is updated on every call to .Write. If the write pushes the length above the chunking threshold, it is then set to -1.

This simplifies the logic and does not cause any of the tests to fail.