Enable closing streamer with error
If a handler is using typhon.Streamer() to stream data to the response
body, there is currently no way for the handler to signal that streaming
terminated unsuccessfully (besides hanging and waiting for a timeout, or
encoding some application-level error into the response body).

This creates a risk of clients not being able to distinguish between
gracefully-or-ungracefully ended streaming HTTP response bodies.

This new CloseWithError() allows the handler to signal an error so that
reads to the reader side of the pipe (passed to http.Response.Body)
return a non-EOF error.