Expose StreamerWriter interface
Streamer() now returns a StreamWriter interface (not io.ReadWriteCloser)
so that the CloseWithError(error) method may be called without casting.

The idiomatic name for this interface and function should be Streamer
and NewStreamer(). However, the function is already named Streamer()
and a package can't contain a function and interface with the same name,
so we need to call it something else. I decided to name the interface
StreamWriter, as writing is the primary use case for library callers.