error-filter: Propogate body read errors
If we fail to read the entire body, we should propogate that error.

This will allow us to stop emitting this I/O error to Sentry.

Don't attempt to recover the partially read bytes - they might contain
sensitive data, and putting them in a terror risks them being propogated
to an unintended target.