errors: export Status2TerrCode
This PR exports the errors.Status2TerrCode function. It is currently unexported, and also unused.

I'd like to export this as it is useful for implementing custom error filters, for example, when using a Typhon client to talk to third-party APIs, it may be necessary to implement custom logic to deserialize errors returned into Terrors. As part of that, it is common to translate the response HTTP status codes into Terror error codes. In our codebase, I can see around 15 re-implementations of this mapping, so I think there is sufficient usage for this change.