Expose RouterPathPatternForRequest
This will allow us to extract the path pattern from a requests context, without having to go via the requests route. When used in client filters, the 'router' will not be set correctly, as the req will be outbound, but the context on the req will have preserved the original inbound requests context values.